shopfloor_mobile: allow extenstion of states on checkout
This change allows for other modules to extend the state of the checkout
scenario.
See the `shopfloor_checkout_package_measurement_mobile` for a use of
this change.
